+++ pybdsim/Field/__init__.py
@@ -11,5 +11,6 @@ from ._Field import Load
 
 from .FieldPlotter import Plot2DXY
 from .FieldPlotter import Plot2DXYBz
+from .FieldPlotter import Plot2DXYFxFyFz
 from .FieldPlotter import Plot3DXY
 from .FieldPlotter import Plot3DXZ